pyproject.toml
[project]
name = "pyplaylisticot"
version = "0.1.0"
description = ""
authors = [
    {name = "Simisticot",email = "simi.pra@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"

[tool.poetry]
package-mode=false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
discord = "^2.3.2"
fastapi = {extras = ["standard"], version = "^0.115.14"}
google-api-python-client = "^2.174.0"
google-auth-oauthlib = "^1.2.2"
google-auth-httplib2 = "^0.2.0"
requests = "^2.32.4"
flask = "^3.1.1"
pytest = "^8.4.1"
